refactor: split Settings context into facade + sub-modules
Extract Settings.QualityProfiles (profile CRUD, cloning, comparison,
export/import), Settings.ServiceConfigs (client/indexer/media server CRUD),
Settings.LibraryPaths (library path CRUD + validation), and
Settings.RuntimeConfig (config settings, typed getters, runtime builders).

Settings.ex becomes a thin facade with 68 defdelegate calls.
No behavior changes, all 134 settings tests pass.
